Don't ignore custom boards.txt path when using a custom ard-parse-boards.
This splits the definition of the path to ard-parse-boards from the arguments
passed to it - most importantly the custom path to the boards.txt which changes
on different platforms. The default value for this hard-coded into
ard-parse-boards won't work in Linux or Windows, so passing the custom value is
critical.

This change also determines the absolute path to ard-parse-boards dynamically,
so the PARSE_BOARDS_PATH value should never really need to be overridden. I
removed the custom values from all of the example Makefiles.