Enable dependency tracking for libraries and core
This makes sure a library or core file gets rebuild when a header file
it depends on is changed, preventing weird surprises when the
application is recompiled with the new header values but the library
isn't.

For most users, this won't be strictly needed, but anyone working on a
library or even the Arduino core, this is must to prevent surprises or
needing a make clean between every build attempt.