Compile .ino and .pde files directly
Before, they were copied to a .cpp file to add the Arduino.h/WProgram.h
include. However, this would cause the compiler error messages to not refer to
the right filename, making it hard to use the compiler output in an editor like
vim to point out errors.

By using gcc's -include option, there is no need to modify the ino/pde
file before compiling. However, we will need to explicitely tell gcc
that the source file is c++, because of the non-standard extensions.