Don't indent variable assignments with tabs
This causes make to consider the line part of the preceding recipe,
causing the variable to remain unset and the preceding recipe to break.
In this case, this was the recipe for user libraries with .c files,
which is uncommon enough for this to go unnoticed.

This was broken in 3bce1d88 (Allow adding extra common dependencies).