Simplified package vs. git detection, fixes #147
Removed ARDMK_FILE as it was in the wrong case and a bit pointless anyway.

Fixed ChipKIT.mk which wouldn't work when packaged and had a lot of
duplicated code (needs checking this still works!)

Removed the git vs. package detection of the arduino-mk subdirectory -
ARDMK_DIR is now calculated based on where the *.mk files are.

Removed ARDMK_PATH as we now check for ard-reset-arduino in the $PATH
when packaged or fall back to it being one level above the *.mk files
in bin/ when checked out of git.

Merged Common.mk back into Arduino.mk as it meant nasty workarounds for
missing routines e.g. dir_if_exists and show_config_variable which
meant that AUTO_ARDUINO_DIR was broken.

So now all you need to set is BOARD_TAG (plus MONITOR_PORT or ISP_PROG
for upload/reset).