Arduino.mk: verify guessed tool paths, use system ones if not found
Currently all tool locations depend guessing the avr tools directory
depending on the location of avr-gcc. While this works perfectly for
seperate installations of those tools, it fails if using
system-installed ones that are ccache enabled (as on Fedora). There,
avr-gcc can be found in /usr/lib64/ccache/avr-gcc, then, it uses the
parent directory and appends bin again, leading it to guess wrong paths
for all tools.

Add a line which verifies that valid tools have been found, and if not,
try to identify proper ones by trying to find them in the PATH
environment variable (or relying on shell features to do so).

This change makes Arduino.mk work on ccache-enabled Fedora
out-of-the-box (with the previous patch for disabling LTO that is).